+++ CHANGELOG.md
@@ -18,6 +18,15 @@
   instances contained in Dart library.
 * Add `LibraryReader.element` to get back to the `LibraryElement` instance.
 * Add `ConstantReader.objectValue` to get back to the `DartObject` instance.
+* Add `ConstantReader.peek` to read a value that returns `null` if not found:
+
+```dart
+// Tries to read the field "token" first, then "_token".
+findTokenField(DartObject o) {
+  final reader = new ConstantReader(o);
+  final token = o.peek('token') ?? o.read('_token');
+}
+```
 
 ## 0.6.1+1